# algoAPI/swagger.yaml
openapi: 3.0.3
info:
  title: Algo API
  description: >-
    Documentation for Algo API. REST API service, offering algorithm execution
    on request. Additionaly, input data can be saved to the service and used
    again. Most paths require a user account, which is managed through the API
    aswell.
  version: '1.0'
  x-logo:
    url: ''
servers:
  - url: 'https://algo-algo.herokuapp.com/'
...
paths:
...
  /user:
    post:
      tags:
        - user
      summary: Creates a user
      description: 'Used for creating a user account, returns a JWT token.'
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
            example:
              username: test
              email: test@test.com
              password: test
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/registration_response'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: Invalid id
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: Error running algorithm
  /user/login:
    post:
      tags:
        - user
      summary: Returns an authentication token for user
      description: 'Used for authentication, returns a JWT token for further authentication.'
      operationId: loginUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/login'
            example:
              username: test
              password: test
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/login_response'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: Invalid id
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: Error running algorithm
  '/user/{user_id}':
    get:
      tags:
        - user
      summary: Returns the user
      description: Used for fetching a user.
      operationId: getUser
      parameters:
        - name: user_id
          in: path
          required: true
          deprecated: false
          style: simple
          explode: false
          schema:
            type: string
          example: 5f3348e14d9b386e025681a1
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/user_response'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: Invalid id
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: Invalid or expired jwt
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: Error running algorithm
      security:
        - JWTAuth: []
    put:
      tags:
        - user
      summary: Updates the user
      description: >-
        Used for updating a user. User can update username, email or password.
        Fields provided in request body will be updated.
      operationId: updateUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/user_update'
            example:
              username: test1
              password: test1
              email: test1
        required: true
      parameters:
        - name: user_id
          in: path
          required: true
          deprecated: false
          style: simple
          explode: false
          schema:
            type: string
          example: 5f3348e14d9b386e025681a1
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: Invalid id
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: Invalid or expired jwt
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: Error running algorithm
      security:
        - JWTAuth: []
    delete:
      tags:
        - user
      summary: Deletes the user
      description: Used for deleting a user.
      operationId: deleteUser
      parameters:
        - name: user_id
          in: path
          required: true
          deprecated: false
          style: simple
          explode: false
          schema:
            type: string
          example: 5f3348e14d9b386e025681a1
      responses:
        '204':
          description: No content
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: Invalid id
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: Invalid or expired jwt
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: Error running algorithm
      security:
        - JWTAuth: []
...
components:
  schemas:
...
    Error:
      required:
        - message
      type: object
      properties:
        message:
          type: string
          description: Error message indicating the issue
      example:
        message: User not found
    User:
      required:
        - email
        - password
        - username
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
      example:
        username: test
        email: test@test.com
        password: test
...
    registration_response:
      required:
        - created
        - email
        - id
        - username
      properties:
        userID:
          type: string
        token:
          type: string
        username:
          type: string
      example:
        userID: 5fe98a1453590b33f7a82d64
        token: >-
          eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE2MDkxNDQzNTYsInVzZXJfaWQiOiI1ZmU5OGExNDUzNTkwYjMzZjdhODJkNjQiLCJ1c2VyX3VzZXJuYW1lIjoidGVzdCJ9.jgLcRffdVVsOzdfGDl09pcQXruEnpolNuyatPp8CflM
        username: test
    login:
      required:
        - password
        - username
      properties:
        username:
          type: string
        password:
          type: string
      example:
        username: test
        password: test
    user_update:
      properties:
        username:
          type: string
        password:
          type: string
        email:
          type: string
      example:
        username: test1
        password: test1
    login_response:
      type: object
      properties:
        userID:
          type: string
        token:
          type: string
        username:
          type: string
      example:
        userID: 5fe98a1453590b33f7a82d64
        token: >-
          eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE2MDkxNDQzNTYsInVzZXJfaWQiOiI1ZmU5OGExNDUzNTkwYjMzZjdhODJkNjQiLCJ1c2VyX3VzZXJuYW1lIjoidGVzdCJ9.jgLcRffdVVsOzdfGDl09pcQXruEnpolNuyatPp8CflM
        username: test
    user_response:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
        created:
          type: string
        updated:
          type: string
      example:
        id: 5fe98c87e557d8082eac6076
        username: test
        email: test@test.com
        created: '2020-12-28T08:43:03+01:00'
        updates: '2020-12-28T08:43:03+01:00'
...
  securitySchemes:
    JWTAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT